+++ src/main/kotlin/csense/idea/kotlin/checked/exceptions/bll/ProjectClassResolutionInterface.kt
@@ -3,8 +3,7 @@ package csense.idea.kotlin.checked.exceptions.bll
 import com.intellij.openapi.project.*
 import csense.idea.base.bll.psiWrapper.`class`.*
 import csense.idea.base.bll.psiWrapper.`class`.operations.*
-import csense.idea.kotlin.checked.exceptions.bll.callthough.*
-import csense.idea.kotlin.checked.exceptions.bll.ignore.*
+import csense.idea.kotlin.checked.exceptions.repo.*
 
 //TODO base module?
 class ProjectClassResolutionInterface(