+++ src/main/kotlin/csense/idea/kotlin/checked/exceptions/repo/IgnoreRepo.kt
@@ -1,8 +1,9 @@
-package csense.idea.kotlin.checked.exceptions.bll.ignore
+package csense.idea.kotlin.checked.exceptions.repo
 
 import com.intellij.openapi.project.*
 import csense.idea.base.bll.kotlin.*
 import csense.idea.kotlin.checked.exceptions.bll.files.*
+import csense.idea.kotlin.checked.exceptions.bll.ignore.*
 
 class IgnoreRepo(
     project: Project