+++ changelog.md
@@ -6,7 +6,7 @@
     - Quick fixes should generally work pretty well now
     - a lot of things that previously did not work (or only partially worked) now works
         - e.g. contracts are now read / interpreted (calls in place)
-    - Highlights simple Throws in kotlin doc (is not fully implemented)
+    - Preliminary support for @throws kotlin doc
     - should respect imports now when generating quick fixe(s)
 - began work on coding throwing functions from kotlin's std lib. (requires a lot of work)
 - Renamed settings to avoid conflicts with older version(s).